# docs/public/api/openapi-3.0.yaml
openapi: 3.0.0
info:
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  version: "1.0"
paths:
...
  "/v1/entities/{id}/stock":
    get:
      security:
        - Bearer: []
      tags:
        - Stock
      summary: Get item stock allocations
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.StockState"
    post:
      security:
        - Bearer: []
      tags:
        - Stock
      summary: Adjust, set, or transfer item stock
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.StockOperationRequest"
        description: Stock operation
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.StockState"
  "/v1/entities/{id}/stock/default":
    post:
      security:
        - Bearer: []
      tags:
        - Stock
      summary: Set the compatibility default stock location
      parameters:
        - description: Entity ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.SetDefaultStockRequest"
        description: Default allocation
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.StockState"
  "/v1/entities/{locationId}/stock-resolution":
    get:
      security:
        - Bearer: []
      tags:
        - Stock
      summary: List stock blocking location deletion
      parameters:
        - description: Location ID
          name: locationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.LocationStockResolutionResult"
    post:
      security:
        - Bearer: []
      tags:
        - Stock
      summary: Resolve stock blocking location deletion
      parameters:
        - description: Location ID
          name: locationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.LocationStockResolutionRequest"
        description: Resolution
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.LocationStockResolutionResult"
...
  "/v1/labelmaker/asset/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Items
      summary: Get Asset label
      parameters:
        - description: Asset ID
          name: id
          in: path
          required: true
          schema:
            type: string
        - description: Print this label, defaults to false
          name: print
          in: query
          schema:
            type: boolean
        - description: Allocation location ID
          name: locationId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: image/png
          content:
            application/json:
              schema:
                type: string
  "/v1/labelmaker/item/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Items
      summary: Get Item label
      parameters:
        - description: Item ID
          name: id
          in: path
          required: true
          schema:
            type: string
        - description: Print this label, defaults to false
          name: print
          in: query
          schema:
            type: boolean
        - description: Allocation location ID
          name: locationId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: image/png
          content:
            application/json:
              schema:
                type: string
  "/v1/labelmaker/location/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Locations
      summary: Get Location label
      parameters:
        - description: Location ID
          name: id
          in: path
          required: true
          schema:
            type: string
        - description: Print this label, defaults to false
          name: print
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: image/png
          content:
            application/json:
              schema:
                type: string
...
  /v1/stock-transactions:
    get:
      security:
        - Bearer: []
      tags:
        - Stock
      summary: Get stock transaction history
      parameters:
        - description: Entity ID
          name: entityId
          in: query
          schema:
            type: string
        - description: Location ID
          name: locationId
          in: query
          schema:
            type: string
        - description: Page
          name: page
          in: query
          schema:
            type: integer
        - description: Page size
          name: pageSize
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.PaginationResult-repo_StockTransaction"
...
servers:
  - url: /api
components:
...
  securitySchemes:
    Bearer:
      description: "\"Type 'Bearer TOKEN' to correctly set the API Key\""
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    repo.EntitySummary:
      type: object
      properties:
        allocatedQuantity:
          type: number
        archived:
          type: boolean
        assetId:
          type: string
          example: "0"
        createdAt:
          type: string
        description:
          type: string
        entityType:
          allOf:
            - $ref: "#/components/schemas/repo.EntityTypeSummary"
          x-omitempty: true
          nullable: true
        id:
          type: string
        imageId:
          type: string
          x-omitempty: true
          nullable: true
        insured:
          type: boolean
        itemCount:
          description: Container-specific (populated when querying locations)
          type: number
        locationCount:
          type: integer
        name:
          type: string
        parent:
          description: Edges
          allOf:
            - $ref: "#/components/schemas/repo.EntitySummary"
          x-omitempty: true
          nullable: true
        purchasePrice:
          type: number
        quantity:
          type: number
        soldDate:
          description: Sale details
          type: string
        tags:
          type: array
          items:
            $ref: "#/components/schemas/repo.TagSummary"
        thumbnailId:
          type: string
          x-omitempty: true
          nullable: true
        updatedAt:
          type: string
...
    repo.EntityTemplateSummary:
      type: object
      properties:
        createdAt:
          type: string
        description:
          type: string
        id:
          type: string
        name:
          type: string
        updatedAt:
          type: string
...
    repo.EntityTypeSummary:
      type: object
      properties:
        createdAt:
          type: string
        defaultTemplate:
          $ref: "#/components/schemas/repo.EntityTemplateSummary"
        defaultTemplateId:
          type: string
        description:
          type: string
        icon:
          type: string
        id:
          type: string
        isLocation:
          type: boolean
        name:
          type: string
        updatedAt:
          type: string
...
    repo.LocationStockConflict:
      type: object
      properties:
        entityId:
          type: string
        entityName:
          type: string
        isDefault:
          type: boolean
        quantity:
          type: number
    repo.LocationStockResolutionRequest:
      type: object
      required:
        - action
        - idempotencyKey
      properties:
        action:
          type: string
          enum:
            - transfer
            - remove
        confirmed:
          type: boolean
        destinationLocationId:
          type: string
          x-omitempty: true
          nullable: true
        idempotencyKey:
          type: string
          maxLength: 255
          minLength: 1
        reason:
          type: string
          maxLength: 1000
        workflow:
          type: string
          maxLength: 100
    repo.LocationStockResolutionResult:
      type: object
      properties:
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/repo.LocationStockConflict"
        itemCount:
          type: integer
        locationId:
          type: string
        totalQuantity:
          type: number
...
    repo.PaginationResult-repo_StockTransaction:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/repo.StockTransaction"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    repo.SetDefaultStockRequest:
      type: object
      properties:
        locationId:
          type: string
          x-omitempty: true
          nullable: true
    repo.StockAllocation:
      type: object
      properties:
        createdAt:
          type: string
        id:
          type: string
        isDefault:
          type: boolean
        itemId:
          type: string
        location:
          allOf:
            - $ref: "#/components/schemas/repo.EntitySummary"
          x-omitempty: true
          nullable: true
        locationId:
          type: string
          nullable: true
        quantity:
          type: number
        updatedAt:
          type: string
    repo.StockLocationSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    repo.StockOperationRequest:
      type: object
      required:
        - idempotencyKey
        - operation
      properties:
        delta:
          type: number
          x-omitempty: true
          nullable: true
        fromLocationId:
          type: string
          x-omitempty: true
          nullable: true
        idempotencyKey:
          type: string
          maxLength: 255
          minLength: 1
        locationId:
          type: string
          x-omitempty: true
          nullable: true
        operation:
          type: string
          enum:
            - adjust
            - set
            - transfer
        quantity:
          type: number
          x-omitempty: true
          nullable: true
        reason:
          type: string
          maxLength: 1000
        setDefault:
          type: boolean
        toLocationId:
          type: string
          x-omitempty: true
          nullable: true
        workflow:
          type: string
          maxLength: 100
    repo.StockState:
      type: object
      properties:
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/repo.StockAllocation"
        defaultLocationId:
          type: string
          nullable: true
        totalQuantity:
          type: number
    repo.StockTransaction:
      type: object
      properties:
        actorId:
          type: string
          x-omitempty: true
          nullable: true
        actorName:
          type: string
        afterTotal:
          type: number
        beforeTotal:
          type: number
        createdAt:
          type: string
        destinationAfter:
          type: number
          x-omitempty: true
          nullable: true
        destinationBefore:
          type: number
          x-omitempty: true
          nullable: true
        destinationLocation:
          allOf:
            - $ref: "#/components/schemas/repo.StockLocationSummary"
          x-omitempty: true
          nullable: true
        destinationLocationId:
          type: string
          x-omitempty: true
          nullable: true
        entityId:
          type: string
        id:
          type: string
        idempotencyKey:
          type: string
        operation:
          type: string
        quantity:
          type: number
        reason:
          type: string
        sourceAfter:
          type: number
          x-omitempty: true
          nullable: true
        sourceBefore:
          type: number
          x-omitempty: true
          nullable: true
        sourceLocation:
          allOf:
            - $ref: "#/components/schemas/repo.StockLocationSummary"
          x-omitempty: true
          nullable: true
        sourceLocationId:
          type: string
          x-omitempty: true
          nullable: true
        workflow:
          type: string
...
    repo.TagSummary:
      type: object
      properties:
        color:
          type: string
        createdAt:
          type: string
        description:
          type: string
        icon:
          type: string
        id:
          type: string
        name:
          type: string
        parentId:
          type: string
          nullable: true
        updatedAt:
          type: string
...